Implement Simple Greedy Register Allocator with Furthest Next Use heuristic
- Add RegisterAllocator class with 18 available physical registers (t0-t6, s1-s11)
- Implement Simple Greedy algorithm: checks if allocated, spilled, or new
- Use Furthest Next Use heuristic for spill selection (furthest next use or unknown)
- Track all uses of each vreg in sorted lists for efficient binary search
- Handle syscall register clobbering (VEVAL uses a0-a3, VCALL uses a0-a5)
- Integrate with RISCVCodeGen for register allocation during code generation
- Add comprehensive Javadoc-style documentation
- Use explicit types throughout (no auto keywords)
- Add test_compilation.cpp with comprehensive register allocation tests:
  * Tests verify no unnecessary spills when registers are available
  * Tests verify minimal spilling when spilling is necessary (7 spills for 25 vars)
  * Tests verify never exceeding 18 allocated registers
  * Edge case tests: empty function, only parameters, stress test (50 vars),
    overlapping live ranges
- Add test_register_allocation.gd for testing register pressure scenarios
- Fix type conversion warnings in codegen, elf_builder, ir_interpreter

+++ src/gdscript/compiler/codegen.cpp
@@ -601,7 +601,7 @@ void CodeGenerator::pop_scope() {
 
 CodeGenerator::Variable* CodeGenerator::find_variable(const std::string& name) {
 	// Search from innermost to outermost scope
-	for (int i = m_scope_stack.size() - 1; i >= 0; i--) {
+	for (int i = static_cast<int>(m_scope_stack.size()) - 1; i >= 0; i--) {
 		auto it = m_scope_stack[i].variables.find(name);
 		if (it != m_scope_stack[i].variables.end()) {
 			return &it->second;
